fix: refactor app structure and improve test setup[#6]
