+++ mindmap_local/services/auth-service/src/__tests__/setup.ts
@@ -18,22 +18,18 @@ afterAll(async () => {
   // Windowsでファイルロックが解除されるまで待つ
   await new Promise(resolve => setTimeout(resolve, 500));
   
-  // テストデータベースファイルを削除
+  // テストデータベースファイルを削除（WALモードの関連ファイルも含む）
   const testDbPath = path.resolve(config.DB_PATH);
-  if (fs.existsSync(testDbPath)) {
-    try {
-      fs.unlinkSync(testDbPath);
-    } catch (error) {
-      // ファイルがまだロックされている場合は次回テスト時に上書きされるので無視
-      // console.warn('Test database cleanup skipped (file locked)');
+  const walFiles = [testDbPath, `${testDbPath}-wal`, `${testDbPath}-shm`];
+  
+  for (const filePath of walFiles) {
+    if (fs.existsSync(filePath)) {
+      try {
+        fs.unlinkSync(filePath);
+      } catch (error) {
+        // ファイルがまだロックされている場合は次回テスト時に上書きされるので無視
+        // console.warn(`Test database cleanup skipped (${path.basename(filePath)} locked)`);
+      }
     }
   }
-});
-
-beforeEach(() => {
-  // 各テスト前にテーブルをクリア
-});
-
-afterEach(() => {
-  // 各テスト後のクリーンアップ
 });
