+++ mindmap_local/services/auth-service/src/index.ts
@@ -1,27 +1,7 @@
-import express from 'express';
-import cors from 'cors';
 import { config } from './config/env';
 import { initDatabase, closeDb } from './db/init';
-import authRoutes from './routes/auth';
+import { createApp } from './app';
 import pkg from '../package.json';
-
-const app = express();
-
-app.use(cors());
-app.use(express.json());
-
-app.use('/auth', authRoutes);
-
-app.get('/health', (_req, res) => {
-	res.status(200).json({
-		status: 'ok',
-		service: 'auth-service',
-		version: pkg.version ?? '0.0.0',
-		uptime: Math.round(process.uptime()),
-		timestamp: new Date().toISOString(),
-		env: config.NODE_ENV,
-	});
-});
 
 try {
 	initDatabase();
@@ -30,6 +10,8 @@ try {
 	console.error('[startup] Database initialization failed:', error);
 	process.exit(1);
 }
+
+const app = createApp();
 
 const server = app.listen(config.PORT, () => {
 	console.log('[startup] Auth service started');
